+++ src/game/cGirls.h
@@ -120,7 +120,6 @@ public:
     std::shared_ptr<sGirl> GetUniqueGirl(bool slave = false, bool catacomb = false, bool arena = false, bool daughter = false, bool isdaughter = false);
     std::shared_ptr<sGirl> GetUniqueYourDaughterGirl(int Human0Monster1 = -1);    // -1 either, 0 human, 1 monster
     std::shared_ptr<sGirl> GetDaughterByName(const std::string& name, bool player_dad);
-    std::shared_ptr<sGirl> CreateDaughter(sGirl& mother, bool player_dad);
 
     static std::string GetHoroscopeName(int month, int day);