+++ src/game/character/traits/cTraitSpec.cpp
@@ -108,8 +108,8 @@ void cTraitSpec::add_modifier(std::string mod, int value) {
 
 cTraitProps::cTraitProps() : cSimpleKeyValue("Setting", "Name", "Value") {
     // inheritance
-    add_setting("inherit:chance", "Inherit Chance", sPercent(50), "Base chance for inheriting this trait.");
-    add_setting("inherit:chance:from_dormant", "Inherit chance (from dormant)", boost::blank(),
+    add_setting(traits::properties::INHERIT_CHANCE, "Inherit Chance", sPercent(50), "Base chance for inheriting this trait.");
+    add_setting(traits::properties::INHERIT_CHANCE_FROM_DORMANT, "Inherit chance (from dormant)", boost::blank(),
                "Chance that this trait will be inherited from a dormant trait.", "inherit:chance");
 
     add_setting("inherit:dormant", "Dormant Chance (inherited)", boost::blank(),
@@ -117,7 +117,7 @@ cTraitProps::cTraitProps() : cSimpleKeyValue("Setting", "Name", "Value") {
     add_setting("inherit:dormant_from_dormant", "Dormant Chance (inherited from dormant)", sPercent(75),
                 "Chance that this trait will be dormant when inherited from dormant.");
     add_setting("inherit:from_acquired", "Transmission Chance", sPercent(0),
-                "Chance that an active aquired trait (such as a disease) is passed on to offspring as an acquired trait.");
+                "Chance that an active acquired trait (such as a disease) is passed on to offspring as an acquired trait.");
 
     // random spawning
     add_setting("random:chance", "Random Chance", sPercent(1), "Chance for this trait to appear in a randomly generated girl if not specified in the template.");